+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Enable transition to DONE through ApplyTaskAction (#785)
+
 ## [0.16.0] - 2022-06-14
 
 ### Changed